Adding return indices to golden and moving the indices check from mpwi test to golden function
